@@ -10,48 +10,22 @@ class JMSSerializer implements SerializerInterface
 {
     use SerializerTrait;
 
-    /**
-     * @var Serializer
-     */
-    private $serializer;
-
-    /**
-     * @var SerializationContext
-     */
-    private $context;
-
-    /**
-     * @param Serializer $serializer
-     * @param SerializationContext|null $context
-     */
-    public function __construct(Serializer $serializer, SerializationContext $context = null)
+    public function __construct(private Serializer $serializer, private ?SerializationContext $context = null)
     {
-        $this->serializer = $serializer;
-        $this->context = $context;
+
     }
 
-    /**
-     * @param $data
-     * @param $format
-     * @return mixed|string
-     */
-    public function serialize($data, $format)
+    public function serialize(mixed $data, string $format): ?string
     {
         return $this->serializer->serialize($data, $format, $this->context);
     }
 
-    /**
-     * @return array
-     */
-    public function getSupportedFormats()
+    public function getSupportedFormats(): array
     {
         return [Format::FORMAT_JSON, Format::FORMAT_XML];
     }
 
-    /**
-     * @return string
-     */
-    public function getDefaultFormat()
+    public function getDefaultFormat(): string
     {
         return Format::FORMAT_JSON;
     }
