@@ -4,5 +4,5 @@ namespace MediaMonks\RestApi\Exception;
 
 interface ExceptionInterface
 {
-    public function toArray();
+    public function toArray(): array;
 }
