@@ -4,5 +4,5 @@ namespace MediaMonks\RestApi\Exception;
 
 interface FieldExceptionInterface
 {
-    public function getFields();
+    public function getFields(): array;
 }
